+++ backend/external_canisters/icp_ledger/c2c_client/src/lib.rs
@@ -1,0 +1,5 @@
+use canister_client::generate_candid_c2c_call;
+use icp_ledger_canister::*;
+
+// Updates
+generate_candid_c2c_call!(transfer);